Add basic combat (#12)
Allow player to attack by clicking the left mouse button. An area is moved in the direction of the mouse, with a constant offset from the attacker, which can be used to detect the targets that are in range of the attack. A cooldown is added for the attacking, making the player continuously attack in constant intervals while the attack button is pressed, being able to change the direction in real time. The body detection for the attack collider is done one frame after moving it, so all targets are added correctly.